+++ frontend/srcs/src/app/app.module.ts
@@ -3,19 +3,21 @@ import { BrowserModule } from '@angular/platform-browser';
 import { HttpClientModule, provideHttpClient, withInterceptors } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { AdministrationComponent,
-         ChatComponent,
-         FriendListComponent,
-         GameComponent,
-         HomeComponent,
-         LoginModule,
-         NavbarComponent,
-         ProfileComponent,
-         SharedAlertModule,
-         SharedAvatarModule,
-         SpectateModule,
-         ChallengeModule, 
-         AdminAlertsComponent} from './modules';
+import {
+  AdministrationComponent,
+  ChatComponent,
+  FriendListComponent,
+  GameComponent,
+  HomeComponent,
+  LoginModule,
+  NavbarComponent,
+  ProfileComponent,
+  SharedAlertModule,
+  SharedAvatarModule,
+  SpectateModule,
+  ChallengeModule,
+  AdminAlertsComponent
+} from './modules';
 import { SocketIoConfig, SocketIoModule } from 'ngx-socket-io';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CookieService } from 'ngx-cookie-service';
@@ -23,28 +25,30 @@ import { environment } from 'src/environments/environment';
 import { ApiInterceptor, ErrorApiInterceptor } from './core';
 import { MySocket } from './services/web-socket.service';
 import { ChannelsCache, ChatsCache, UsersCache } from './cache';
-import { AvatarComponent,
-         BlockSidebarComponent,
-         ChannelAvatarComponent,
-         ChannelHeaderComponent,
-         ChannelInfoComponent,
-         ChannelManagementActionsComponent,
-         ChannelManagementComponent,
-         ChannelManagementUsersComponent,
-         ChannelSidebarComponent,
-         ChannelWindowComponent,
-         ChannelWindowMessageComponent,
-         ChatInfoComponent,
-         ChatSidebarComponent,
-         ChatWindowComponent,
-         ChatWindowMessageComponent,
-         FriendSidebarComponent,
-         AdminstrationUsersListComponent,
-         AdminstrationChannelsComponent,
-         AdministrationChannelManagementUsersComponent,
-         AdministrationChannelManagementActionsComponent,
-         AdministrationUsersManagementActionsComponent, 
-        AdministrationChannelWindowComponent} from './components';
+import {
+  AvatarComponent,
+  BlockSidebarComponent,
+  ChannelAvatarComponent,
+  ChannelHeaderComponent,
+  ChannelInfoComponent,
+  ChannelManagementActionsComponent,
+  ChannelManagementComponent,
+  ChannelManagementUsersComponent,
+  ChannelSidebarComponent,
+  ChannelWindowComponent,
+  ChannelWindowMessageComponent,
+  ChatInfoComponent,
+  ChatSidebarComponent,
+  ChatWindowComponent,
+  ChatWindowMessageComponent,
+  FriendSidebarComponent,
+  AdminstrationUsersListComponent,
+  AdminstrationChannelsComponent,
+  AdministrationChannelManagementUsersComponent,
+  AdministrationChannelManagementActionsComponent,
+  AdministrationUsersManagementActionsComponent,
+  AdministrationChannelWindowComponent
+} from './components';
 import { DateMutations } from './utils';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { TagModule } from 'primeng/tag';
@@ -57,6 +61,9 @@ import { TooltipModule } from 'primeng/tooltip';
 import { RankTableModule } from './modules/rank-table/rank-table.module'
 import { TruncatePipe } from './pipes/truncate.pipe';
 import { ChannelAvatarEditorComponent } from './components/avatar/channel-avatar-editor/channel-avatar-editor.component';
+import { OffCanvasFullStats } from './modules/fullstats/fullstats.modulle';
+//import { PongTVComponent } from './modules/pongtv/pongtv.component';
+
 import { NgbDatepickerModule } from '@ng-bootstrap/ng-bootstrap';
 
 const config: SocketIoConfig = { url: environment.apiUrl };
@@ -121,7 +128,8 @@ const config: SocketIoConfig = { url: environment.apiUrl };
     TooltipModule,
     RankTableModule,
     ChallengeModule,
-    NgbDatepickerModule,    
+    OffCanvasFullStats,
+    NgbDatepickerModule,
 
   ],
   providers: [