+++ frontend/srcs/src/app/models/fullstats/fullstats.model.ts
@@ -1,0 +1,14 @@
+export interface FullStats {
+  userId: string
+  login: string
+  gamesWin: number
+  gamesLose: number
+  gamesDraw: number
+  goalsFavor: number
+  goalsAgainst: number
+  disconect: number
+  points: number
+  avatar: string
+  position: number
+  total: number
+}