+++ frontend/srcs/src/app/modules/fullstats/fullstats.modulle.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { OffCanvasFullStatsComponent } from './offcanvas/offcanvas.component';
+
+@NgModule({
+  declarations: [OffCanvasFullStatsComponent],
+  imports: [CommonModule],
+  exports: [OffCanvasFullStatsComponent],
+})
+export class OffCanvasFullStats { }