+++ frontend/srcs/src/app/modules/rank-table/rank-table.module.ts
@@ -3,17 +3,11 @@ import { CommonModule } from '@angular/common';
 import { RankTableComponent } from './rank-table/rank-table.component'
 import { FormsModule } from '@angular/forms';
 import { FilterPipe } from 'src/app/pipes/filter.pipe';
-
-//@Component({
-//  selector: 'app-rank-module-table',
-//  templateUrl: './rank-table.module.html',
-//  //  styleUrl: './rank-table.component.scss',
-//})
-//export class RankTableComponent { }
+import { OffCanvasFullStats } from '../fullstats/fullstats.modulle';
 
 @NgModule({
   declarations: [RankTableComponent, FilterPipe],
-  imports: [CommonModule, FormsModule],
+  imports: [CommonModule, FormsModule, OffCanvasFullStats],
   exports: [RankTableComponent],
 })
 export class RankTableModule { }